animace

+++ app/api/recipes/route.ts
@@ -1,6 +1,7 @@
 import { connectDatabase } from "@/lib/mongodb";
 import Recipe from "@/models/recipe";
 import { NextResponse } from "next/server";
+import User from "@/models/user";
 
 export async function POST(req: Request) {
   const {
@@ -35,7 +36,35 @@ export async function GET() {
 
 export async function DELETE(req: any) {
   const id = req.nextUrl.searchParams.get("id");
+  if (!id) {
+    throw new Error("Missing recipe id in request");
+  }
+
   await connectDatabase();
-  await Recipe.findByIdAndDelete(id);
+
+  // Smazání receptu
+  const deletedRecipe = await Recipe.findByIdAndDelete(id);
+  if (!deletedRecipe) {
+    throw new Error(`Recipe not found with id ${id}`);
+  }
+
+  const users = await User.find({ favoriteRecipes: id });
+
+  for (const user of users) {
+    if (!user.favoriteRecipes) {
+      throw new Error(`User ${user._id} does not have favoriteRecipes`);
+    }
+
+    const index = user.favoriteRecipes.indexOf(id);
+    if (index === -1) {
+      throw new Error(
+        `Recipe with id ${id} is not in user ${user._id}.favoriteRecipes`
+      );
+    }
+
+    user.favoriteRecipes.splice(index, 1);
+    await user.save();
+  }
+
   return NextResponse.json({ message: "Recipe deleted" }, { status: 200 });
 }
